Add CLI for EPUB processing and audio generation
